+++ backend/src/services/logEntry/utils/mentions/reconcileFileMentions.ts
@@ -3,16 +3,12 @@
 // and is legally attributed to the Department for Business and Trade (UK) as the governing entity.
 
 import { type LogEntry } from 'common/LogEntry';
+import { LogEntryContent } from 'common/LogEntryContent';
 import { getMentionsOfType } from './utils';
 import { FileNameMapping } from '../types';
 
-export function reconcileFileMentions(entry: LogEntry, entryId: string, namesMap: FileNameMapping[]) {
-  const { content } = entry;
-  if (!content.json) {
-    return;
-  }
-
-  const fileMentions = getMentionsOfType(entry, 'File');
+function getMentions(content: LogEntryContent, entryId: string, namesMap: FileNameMapping[]) {
+  const fileMentions = getMentionsOfType(content, 'File');
   if (!fileMentions.length) {
     return;
   }
@@ -21,7 +17,19 @@ export function reconcileFileMentions(entry: LogEntry, entryId: string, namesMap
     if (mention.id.startsWith('this::')) {
       const fileName = mention.id.split('::')[1];
       const mapping = namesMap.find((map) => map.originalname === fileName);
+      // eslint-disable-next-line no-param-reassign
       content.json = content.json.replace(mention.id, `${entryId}::${mapping?.storedName || fileName}`);
     }
   });
 }
+
+export function reconcileFileMentionsFromLogContent(entry: LogEntry, entryId: string, namesMap: FileNameMapping[]) {
+  const { content } = entry;
+  if (!content.json) {
+    return;
+  }
+
+  getMentions(content, entryId, namesMap);
+}
+
+