+++ backend/src/services/logEntry/utils/select/mentionsUser.ts
@@ -6,7 +6,7 @@
 import { select } from '../../../../ia';
 import { ns } from '../../../../rdfutil';
 
-export function mentionsUser(incidentId: string) {
+export function mentionsUserInLogContent(incidentId: string) {
   return select({
     clause: [
       [ns.data(incidentId), ns.lisa.hasLogEntry, '?entryId'],